Add creation date file NCBI taxonomy download

+++ Snakefile
@@ -225,33 +225,40 @@ rule DownloadNCBITaxonomy:
 		donefile = "{workingdirectory}/taxdownload.done.txt"
 	shell:
 		"""
-		curl -R ftp://ftp.ncbi.nih.gov/pub/taxonomy/taxdump.tar.gz.md5 --output {input.taxdir}/taxdump.tar.gz.md5
-		if [ -f {input.taxdir}/names.dmp ]; then
-			if [ {input.taxdir}/taxdump.tar.gz.md5 -nt {input.taxdir}/names.dmp ]; then
+		today=$(date +%d-%m-%Y)
+		if [ ! -f {datadir}/taxonomy/date.$today.txt ]; then
+			if [ -f {datadir}/taxonomy/date.*txt ]; then
+				rm {datadir}/taxonomy/date.*txt
+			fi
+			touch {datadir}/taxonomy/date.$today.txt
+			curl -R ftp://ftp.ncbi.nih.gov/pub/taxonomy/taxdump.tar.gz.md5 --output {input.taxdir}/taxdump.tar.gz.md5
+			if [ -f {input.taxdir}/names.dmp ]; then
+				if [ {input.taxdir}/taxdump.tar.gz.md5 -nt {input.taxdir}/names.dmp ]; then
+					curl -R ftp://ftp.ncbi.nih.gov/pub/taxonomy/taxdump.tar.gz --output taxdump.tar.gz
+					tar -C {input.taxdir} -xzf taxdump.tar.gz names.dmp nodes.dmp
+					rm taxdump.tar.gz
+				fi
+			else
 				curl -R ftp://ftp.ncbi.nih.gov/pub/taxonomy/taxdump.tar.gz --output taxdump.tar.gz
 				tar -C {input.taxdir} -xzf taxdump.tar.gz names.dmp nodes.dmp
 				rm taxdump.tar.gz
 			fi
-		else
-			curl -R ftp://ftp.ncbi.nih.gov/pub/taxonomy/taxdump.tar.gz --output taxdump.tar.gz
-			tar -C {input.taxdir} -xzf taxdump.tar.gz names.dmp nodes.dmp
-			rm taxdump.tar.gz
-		fi
-		curl -R ftp.ncbi.nih.gov/pub/taxonomy/accession2taxid/nucl_gb.accession2taxid.gz.md5 --output {input.taxdir}/nucl_gb.accession2taxid.gz.md5
-		if [ -f {input.taxdir}/nucl_gb.accession2taxid ]; then
-			if [ {input.taxdir}/nucl_gb.accession2taxid.gz.md5 -nt {input.taxdir}/nucl_gb.accession2taxid ]; then
+			curl -R ftp.ncbi.nih.gov/pub/taxonomy/accession2taxid/nucl_gb.accession2taxid.gz.md5 --output {input.taxdir}/nucl_gb.accession2taxid.gz.md5
+			if [ -f {input.taxdir}/nucl_gb.accession2taxid ]; then
+				if [ {input.taxdir}/nucl_gb.accession2taxid.gz.md5 -nt {input.taxdir}/nucl_gb.accession2taxid ]; then
+					curl -R ftp.ncbi.nih.gov/pub/taxonomy/accession2taxid/nucl_gb.accession2taxid.gz --output nucl_gb.accession2taxid.gz
+					curl -R ftp.ncbi.nih.gov/pub/taxonomy/accession2taxid/nucl_wgs.accession2taxid.gz --output nucl_wgs.accession2taxid.gz
+					gzip -dc nucl_gb.accession2taxid.gz > {input.taxdir}/nucl_gb.accession2taxid
+					gzip -dc nucl_wgs.accession2taxid.gz > {input.taxdir}/nucl_wgs.accession2taxid
+					rm nucl_gb.accession2taxid.gz nucl_wgs.accession2taxid.gz
+				fi
+			else
 				curl -R ftp.ncbi.nih.gov/pub/taxonomy/accession2taxid/nucl_gb.accession2taxid.gz --output nucl_gb.accession2taxid.gz
 				curl -R ftp.ncbi.nih.gov/pub/taxonomy/accession2taxid/nucl_wgs.accession2taxid.gz --output nucl_wgs.accession2taxid.gz
 				gzip -dc nucl_gb.accession2taxid.gz > {input.taxdir}/nucl_gb.accession2taxid
 				gzip -dc nucl_wgs.accession2taxid.gz > {input.taxdir}/nucl_wgs.accession2taxid
-				rm nucl_gb.accession2taxid.gz nucl_wgs.accession2taxid.gz
+				rm nucl_gb.accession2taxid.gz nucl_wgs.accession2taxid.gz			
 			fi
-		else
-			curl -R ftp.ncbi.nih.gov/pub/taxonomy/accession2taxid/nucl_gb.accession2taxid.gz --output nucl_gb.accession2taxid.gz
-			curl -R ftp.ncbi.nih.gov/pub/taxonomy/accession2taxid/nucl_wgs.accession2taxid.gz --output nucl_wgs.accession2taxid.gz
-			gzip -dc nucl_gb.accession2taxid.gz > {input.taxdir}/nucl_gb.accession2taxid
-			gzip -dc nucl_wgs.accession2taxid.gz > {input.taxdir}/nucl_wgs.accession2taxid
-			rm nucl_gb.accession2taxid.gz nucl_wgs.accession2taxid.gz			
 		fi
 		touch {output.donefile}
 		"""
@@ -1060,6 +1067,7 @@ rule concatenate_reads:
 		fi
 		"""
 
+'''
 def aggregate_readsetslist(wildcards):
 	checkpoint_output=checkpoints.GetGenera.get(**wildcards).output[0]
 	return expand ("{workingdirectory}/{genus}/{genus}.readsids.txt", workingdirectory=config["workingdirectory"], genus=glob_wildcards(os.path.join(checkpoint_output, 'genus.{genus}.txt')).genus)
@@ -1086,6 +1094,7 @@ rule concatenate_readlist:
 			cp {reads} {output.target}
 		fi
 		"""
+'''
 
 def aggregate_readsets_putative(wildcards):
 	checkpoint_output=checkpoints.GetGenera.get(**wildcards).output[0]
